Change: data structure implementations were switched to structs from classes.
This is due to the fact that the app is based on call by references and heap allocations which are the basis of linked lists. As such, after considering a multitude of  strategies including classes, private members etc., it was decided to use aggregates instead. What's the point of making a pointer member private? We'll then need getters and setters but those won't provide the protection that getters and setters normally provide.